Created Custom Chip to Fix Double Element Remove Bug
Initially, i used the PrimeVue Chips component's remove callback to remove an item from my own list. However, that caused a double remove on the dom element but not in the list iteself.
Turns out that it is probably due to some collision of remove functionality.

When i implement my own custom chip function, there was no base remove functionality of the element itself and the remove was handled by my own remove function - and the chips are populated based off that list.

In my own custom function, if i implemented a dom removal function, i would then be able to replicate the same error as the default PrimeVue component.

Pretty interesting.